pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "netbox-zpl-labels"
version = "0.1.0"
description = "NetBox plugin for generating and printing ZPL labels for cables"
readme = "README.md"
license = {text = "Apache-2.0"}
authors = [
    {name = "Elvis Smeers", email = "elvis@smeers.nl"}
]
keywords = ["netbox", "plugin", "zpl", "labels", "zebra", "printer", "cables"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: Django",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: System :: Networking",
]
requires-python = ">=3.12"
dependencies = []

[project.optional-dependencies]
preview = [
    "requests>=2.28",
    "Pillow>=9.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-django>=4.5",
    "ruff>=0.1.0",
    "mypy>=1.0",
    "pre-commit>=3.0",
]
all = [
    "netbox-zpl-labels[preview,dev]",
]

[project.urls]
Homepage = "https://github.com/ctrl-alt-automate/netbox-zpl-labels"
Documentation = "https://github.com/ctrl-alt-automate/netbox-zpl-labels#readme"
Repository = "https://github.com/ctrl-alt-automate/netbox-zpl-labels.git"
Issues = "https://github.com/ctrl-alt-automate/netbox-zpl-labels/issues"

[tool.setuptools.packages.find]
include = ["netbox_zpl_labels*"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "B",     # flake8-bugbear
    "C4",    # flake8-comprehensions
    "UP",    # pyupgrade
]
ignore = [
    "E501",  # Line too long (handled by formatter)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["E501"]
"**/migrations/*" = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["netbox_zpl_labels"]
known-third-party = ["django", "rest_framework", "netbox"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = [
    "migrations/",
    "tests/",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "netbox.settings"
python_files = ["test_*.py"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["netbox_zpl_labels"]
omit = [
    "*/migrations/*",
    "*/tests/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
